[ATR-483] refactor: subscribe 폴더 이동
- archive에서 introduction으로 이동
