+++ src/main/java/run/attraction/api/v1/introduction/repository/SubscribeRepository.java
@@ -1,11 +1,11 @@
-package run.attraction.api.v1.archive.repository;
+package run.attraction.api.v1.introduction.repository;
 
 import io.lettuce.core.dynamic.annotation.Param;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-import run.attraction.api.v1.archive.Subscribe;
+import run.attraction.api.v1.introduction.Subscribe;
 
 public interface SubscribeRepository extends JpaRepository<Subscribe, Long> {
   Optional<Subscribe> findByUserEmail(String userEmail);
