+++ src/main/java/run/attraction/api/v1/mypage/service/archive/newsletter/MypageNewsletterServiceImpl.java
@@ -5,8 +5,8 @@ import java.util.List;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
-import run.attraction.api.v1.archive.Subscribe;
-import run.attraction.api.v1.archive.repository.SubscribeRepository;
+import run.attraction.api.v1.introduction.Subscribe;
+import run.attraction.api.v1.introduction.repository.SubscribeRepository;
 import run.attraction.api.v1.introduction.Newsletter;
 import run.attraction.api.v1.introduction.repository.NewsletterRepository;
 import run.attraction.api.v1.mypage.service.dto.archive.newsletter.MypageNewsletterDetail;
