+++ src/test/java/run/attraction/api/v1/mypage/service/archive/MypageNewsletterSerivceTest.java
@@ -11,9 +11,8 @@ import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
-import org.springframework.test.context.TestPropertySource;
-import run.attraction.api.v1.archive.Subscribe;
-import run.attraction.api.v1.archive.repository.SubscribeRepository;
+import run.attraction.api.v1.introduction.Subscribe;
+import run.attraction.api.v1.introduction.repository.SubscribeRepository;
 import run.attraction.api.v1.introduction.Newsletter;
 import run.attraction.api.v1.introduction.repository.NewsletterRepository;
 import run.attraction.api.v1.mypage.service.archive.newsletter.MypageNewsletterServiceImpl;
